Crear un vuelo desde admin
falta agregar una lista de aviones para selecionar

+++ src/com/company/Vuelo.java
@@ -162,7 +162,7 @@ public class Vuelo {
         return cantidadAcompanantes;
     }
 
-    public void setCantidadAcompanantes(short cantidadAcompanantes) {
+    public void setCantidadAcompanantes(int cantidadAcompanantes) {
         this.cantidadAcompanantes = cantidadAcompanantes;
     }
